behavior.mpd - update confirmation, quickplay, youtube
-mpd update now waits until the update completes and announces to
the channel as such
-quickPlay was added, which can be used to pause the playlsit,
add a file, skip to a given time, play for a certian amount of time
and then return back to the playlist.
-mpd youtube <url> works!  It system-calls out to ssh which
executes youtube-dl on the mpd server (mpd.local).  Then it updates
the db and adds the new file to the bottom of the playlist.  It's
slow.